feat: migrate PlayerSystem business logic to pure test layer
- Extract 15 pure business logic tests from PlayerSystem.test.ts
- Create src/test/pure/business/PlayerSystem.test.ts with core functionality:
  - Input handling and hasMovedOnce state management
  - Auto jump system with physics calculations
  - Trail system with length limits and consistency
  - Speed clamping logic
  - Reset functionality and timer management
  - Input integration with movement processing
- Retain framework-dependent landing prediction tests in original file
- Maintain 100% test coverage for migrated business logic
- Pure logic layer now: 40 tests in 477ms execution time
- All quality gates pass: format, lint, typecheck, coverage
